@@ -15,7 +15,7 @@
 //
 module hopsworks.ai/rdrs
 
-go 1.17
+go 1.20
 
 require (
 	github.com/gin-contrib/sse v0.1.0 // indirect
@@ -25,7 +25,6 @@ require (
 	github.com/go-playground/validator/v10 v10.11.2 // indirect
 	github.com/go-sql-driver/mysql v1.7.0
 	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/ianlancetaylor/cgosymbolizer v0.0.0-20220405231054-a1ae3e4bba26
 	github.com/json-iterator/go v1.1.12 // indirect
 	github.com/leodido/go-urn v1.2.2 // indirect
 	github.com/mattn/go-isatty v0.0.17 // indirect
@@ -36,7 +35,6 @@ require (
 	golang.org/x/crypto v0.6.0 // indirect
 	golang.org/x/sys v0.5.0 // indirect
 	gopkg.in/natefinch/lumberjack.v2 v2.2.1
-	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
 
 require (
@@ -45,10 +43,10 @@ require (
 )
 
 require (
-	github.com/BurntSushi/toml v1.1.0 // indirect
 	github.com/bytedance/sonic v1.8.3 // indirect
 	github.com/chenzhuoyu/base64x v0.0.0-20221115062448-fe3a3abad311 // indirect
 	github.com/goccy/go-json v0.10.0 // indirect
+	github.com/ianlancetaylor/cgosymbolizer v0.0.0-20230328201059-365e72989107
 	github.com/klauspost/cpuid/v2 v2.2.4 // indirect
 	github.com/pelletier/go-toml/v2 v2.0.7 // indirect
 	github.com/twitchyliquid64/golang-asm v0.15.1 // indirect
@@ -57,5 +55,4 @@ require (
 	golang.org/x/text v0.7.0 // indirect
 	google.golang.org/genproto v0.0.0-20230301171018-9ab4bdc49ad5 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
-	zappem.net/pub/debug/xxd v0.5.0 // indirect
 )
