@@ -50,12 +50,13 @@ extern "C" {
 #define PK_REQ_OP_TYPE_IDX   0
 #define PK_REQ_CAPACITY_IDX  1
 #define PK_REQ_LENGTH_IDX    2
-#define PK_REQ_DB_IDX        3
-#define PK_REQ_TABLE_IDX     4
-#define PK_REQ_PK_COLS_IDX   5
-#define PK_REQ_READ_COLS_IDX 6
-#define PK_REQ_OP_ID_IDX     7
-#define PK_REQ_HEADER_END    32
+#define PK_REQ_FLAGS_IDX     3
+#define PK_REQ_DB_IDX        4
+#define PK_REQ_TABLE_IDX     5
+#define PK_REQ_PK_COLS_IDX   6
+#define PK_REQ_READ_COLS_IDX 7
+#define PK_REQ_OP_ID_IDX     8
+#define PK_REQ_HEADER_END    36 // NOTE: Update this when you add / remove fields from  header 
 
 // Primary Key Read Response Header Indexes
 #define PK_RESP_OP_TYPE_IDX   0
@@ -66,7 +67,7 @@ extern "C" {
 #define PK_RESP_TABLE_IDX     5
 #define PK_RESP_COLS_IDX      6
 #define PK_RESP_OP_ID_IDX     7
-#define PK_RESP_HEADER_END    32
+#define PK_RESP_HEADER_END    32 // NOTE: Update this when you add / remove fields from  header 
 
 // Hopsworks
 #define API_KEY_PREFIX_SIZE           45 + 1
