@@ -18,6 +18,7 @@
  */
 
 #include "src/rdrs-dal.h"
+#include <unistd.h>
 #include <mgmapi.h>
 #include <cstdlib>
 #include <cstring>
@@ -25,17 +26,21 @@
 #include <iostream>
 #include <iterator>
 #include <sstream>
+#include <my_base.h>
 #include <NdbApi.hpp>
+#include <storage/ndb/include/ndb_global.h>
 #include "src/error-strings.h"
 #include "src/logger.hpp"
 #include "src/db-operations/pk/pkr-operation.hpp"
 #include "src/status.hpp"
+#include "src/retry_handler.hpp"
 #include "src/ndb_object_pool.hpp"
 #include "src/db-operations/pk/common.hpp"
 
-int GetAvailableAPINode(const char *connection_string);
-
 Ndb_cluster_connection *ndb_connection;
+Uint32 OP_RETRY_COUNT               = 3;
+Uint32 OP_RETRY_INITIAL_DELAY_IN_MS = 500;
+Uint32 OP_RETRY_JITTER_IN_MS        = 100;
 
 /**
  * Initialize NDB connection
@@ -44,36 +49,35 @@ Ndb_cluster_connection *ndb_connection;
  * connect to
  * @return status
  */
-RS_Status init(const char *connection_string, unsigned int find_available_node_id) {
+RS_Status init(const char *connection_string, unsigned int connection_pool_size,
+               unsigned int *node_ids, unsigned int node_ids_len, unsigned int connection_retries,
+               unsigned int connection_retry_delay_in_sec) {
+
+  // disable buffered stdout
+  setbuf(stdout, NULL);
+
+  require(node_ids_len == 1);
+  require(connection_pool_size == 1);
   int retCode = 0;
-  DEBUG(std::string("Connecting to ") + connection_string);
+  INFO(std::string("Connecting to ") + connection_string);
 
   retCode = ndb_init();
   if (retCode != 0) {
     return RS_SERVER_ERROR(ERROR_001 + std::string(" RetCode: ") + std::to_string(retCode));
   }
 
-  int node_id = -1;
-  if (find_available_node_id != 0) {
-    node_id = GetAvailableAPINode(connection_string);
-    if (node_id == -1) {
-      return RS_SERVER_ERROR(ERROR_024);
-    }
-  }
-
-  if (node_id != -1) {
-    ndb_connection = new Ndb_cluster_connection(connection_string, node_id);
-  } else {
-    ndb_connection = new Ndb_cluster_connection(connection_string);
-  }
-  retCode = ndb_connection->connect();
+  ndb_connection = new Ndb_cluster_connection(connection_string, node_ids[0]);
+  retCode        = ndb_connection->connect(connection_retries, connection_retry_delay_in_sec, 0);
   if (retCode != 0) {
     return RS_SERVER_ERROR(ERROR_002 + std::string(" RetCode: ") + std::to_string(retCode));
   }
 
   retCode = ndb_connection->wait_until_ready(30, 0);
   if (retCode != 0) {
-    return RS_SERVER_ERROR(ERROR_003 + std::string(" RetCode: ") + std::to_string(retCode));
+    return RS_SERVER_ERROR(
+        ERROR_003 + std::string(" RetCode: ") + std::to_string(retCode) +
+        std::string(" Lastest Error: ") + std::to_string(ndb_connection->get_latest_error()) +
+        std::string(" Lastest Error Msg: ") + std::string(ndb_connection->get_latest_error_msg()));
   }
 
   // Initialize NDB Object Pool
@@ -83,11 +87,21 @@ RS_Status init(const char *connection_string, unsigned int find_available_node_i
   return RS_OK;
 }
 
+RS_Status set_op_retry_props(const unsigned int retry_cont, const unsigned int rety_initial_delay,
+                             const unsigned int jitter) {
+  OP_RETRY_COUNT               = retry_cont;
+  OP_RETRY_INITIAL_DELAY_IN_MS = rety_initial_delay;
+  OP_RETRY_JITTER_IN_MS        = jitter;
+
+  return RS_OK;
+}
+
 RS_Status shutdown_connection() {
   try {
     // ndb_end(0); // causes seg faults when called repeated from unit tests*/
     NdbObjectPool::GetInstance()->Close();
     delete ndb_connection;
+    ndb_connection = nullptr;
   } catch (...) {
     WARN("Exception in Shutdown");
   }
@@ -97,7 +111,7 @@ RS_Status shutdown_connection() {
 /**
  * Closes a NDB Object
  *
- * @param[int] ndb_object
+ * @param[in] ndb_object
  *
  * @return status
  */
@@ -106,6 +120,7 @@ RS_Status closeNDBObject(Ndb *ndb_object) {
   return RS_OK;
 }
 
+
 RS_Status pk_read(RS_Buffer *reqBuff, RS_Buffer *respBuff) {
   Ndb *ndb_object  = nullptr;
   RS_Status status = NdbObjectPool::GetInstance()->GetNdbObject(ndb_connection, &ndb_object);
@@ -113,15 +128,15 @@ RS_Status pk_read(RS_Buffer *reqBuff, RS_Buffer *respBuff) {
     return status;
   }
 
-  PKROperation pkread(reqBuff, respBuff, ndb_object);
+  /* clang-format off */
+  RETRY_HANDLER(
+      PKROperation pkread(reqBuff, respBuff, ndb_object);
+      status = pkread.PerformOperation();
+  )
+  /* clang-format on */
 
-  status = pkread.PerformOperation();
   closeNDBObject(ndb_object);
-  if (status.http_code != SUCCESS) {
-    return status;
-  }
-
-  return RS_OK;
+  return status;
 }
 
 /**
@@ -135,19 +150,19 @@ RS_Status pk_batch_read(unsigned int no_req, RS_Buffer *req_buffs, RS_Buffer *re
     return status;
   }
 
-  PKROperation pkread(no_req, req_buffs, resp_buffs, ndb_object);
+  /* clang-format off */
+  RETRY_HANDLER(
+      PKROperation pkread(no_req, req_buffs, resp_buffs, ndb_object);
+      status = pkread.PerformOperation();
+  )
+  /* clang-format on */
 
-  status = pkread.PerformOperation();
   closeNDBObject(ndb_object);
-  if (status.http_code != SUCCESS) {
-    return status;
-  }
-
-  return RS_OK;
+  return status;
 }
 
 /**
- * Deallocate pointer array
+ * Returns statistis about RonDB connection
  */
 RS_Status get_rondb_stats(RonDB_Stats *stats) {
   RonDB_Stats ret              = NdbObjectPool::GetInstance()->GetStats();
@@ -157,65 +172,6 @@ RS_Status get_rondb_stats(RonDB_Stats *stats) {
   stats->ndb_objects_available = ret.ndb_objects_available;
 
   return RS_OK;
-}
-
-static int LastConnectedInodeID = -1;
-/*
- * NDB API does not support gracefull disconnection form the
- * cluster. All disconnections are treated as failures. When
- * you disconnect, the API node is not able to accept new
- * connections until the filure recovery has completed for
- * the API node. This can take upto ~5 sec, slowing down
- * unit tests which start/stop the NDB API multiple times.
- * This function returns next available API node to connect to.
- */
-int GetAvailableAPINode(const char *connection_string) {
-  NdbMgmHandle h;
-
-  h = ndb_mgm_create_handle();
-  if (h == 0) {
-    PANIC("Failed to create mgm handle");
-    return -1;
-  }
-
-  if (ndb_mgm_set_connectstring(h, connection_string) == -1) {
-    PANIC("Failed set mgm connect string");
-    return -1;
-  }
-
-  if (ndb_mgm_connect(h, 0, 0, 0)) {
-    PANIC("Failed to connect to mgm node");
-    return -1;
-  }
-
-  // look for api nodes only
-  ndb_mgm_node_type node_types[2]   = {NDB_MGM_NODE_TYPE_API, NDB_MGM_NODE_TYPE_UNKNOWN};
-  struct ndb_mgm_cluster_state *ret = ndb_mgm_get_status2(h, node_types);
-
-  if (ret->no_of_nodes > 1) {
-    int max_node_id = ret->node_states[0].node_id;
-    for (int i = 1; i < ret->no_of_nodes; i++) {
-      if (ret->node_states[i].node_id > max_node_id) {
-        max_node_id = ret->node_states[i].node_id;
-      }
-    }
-
-    if (LastConnectedInodeID == max_node_id) {
-      LastConnectedInodeID = -1;
-    }
-
-    for (int i = 0; i < ret->no_of_nodes; i++) {
-      if (ret->node_states[i].node_id > LastConnectedInodeID &&
-          ret->node_states[i].node_status == NDB_MGM_NODE_STATUS_NO_CONTACT) {
-        LastConnectedInodeID = ret->node_states[i].node_id;
-        free(ret);
-        return LastConnectedInodeID;
-      }
-    }
-  }
-
-  free(ret);
-  return -1;
 }
 
 /**
