@@ -1,6 +1,6 @@
 /*
  * This file is part of the RonDB REST API Server
- * Copyright (c) 2022 Hopsworks AB
+ * Copyright (c) 2023 Hopsworks AB
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -18,6 +18,7 @@ package batchpkread
 
 import (
 	"errors"
+	"fmt"
 	"net/http"
 
 	"hopsworks.ai/rdrs/internal/config"
@@ -29,17 +30,18 @@ import (
 )
 
 type Handler struct {
-	heap *heap.Heap
+	heap        *heap.Heap
+	apiKeyCache apikey.Cache
 }
 
-func New(heap *heap.Heap) Handler {
-	return Handler{heap}
+func New(heap *heap.Heap, apiKeyCache apikey.Cache) Handler {
+	return Handler{heap, apiKeyCache}
 }
 
 // TODO: This might have to be instantiated properly
 var pkReadHandler pkread.Handler
 
-func (h Handler) Validate(request interface{}) error {
+func (h *Handler) Validate(request interface{}) error {
 	pkOperations := request.(*[]*api.PKReadParams)
 	if pkOperations == nil {
 		return errors.New("operations is missing in payload")
@@ -55,9 +57,9 @@ func (h Handler) Validate(request interface{}) error {
 	return nil
 }
 
-func (h Handler) Authenticate(apiKey *string, request interface{}) error {
+func (h *Handler) Authenticate(apiKey *string, request interface{}) error {
 	conf := config.GetAll()
-	if !conf.Security.UseHopsworksAPIKeys {
+	if !conf.Security.APIKey.UseHopsworksAPIKeys {
 		return nil
 	}
 
@@ -74,10 +76,14 @@ func (h Handler) Authenticate(apiKey *string, request interface{}) error {
 		dbArr = append(dbArr, &dbKey)
 	}
 
-	return apikey.ValidateAPIKey(apiKey, dbArr...)
+	err := h.apiKeyCache.ValidateAPIKey(apiKey, dbArr...)
+	if err != nil {
+		fmt.Printf("Validation failed. key %s, error  %v \n", *apiKey, err)
+	}
+	return err
 }
 
-func (h Handler) Execute(request interface{}, response interface{}) (int, error) {
+func (h *Handler) Execute(request interface{}, response interface{}) (int, error) {
 	pkOperations := request.(*[]*api.PKReadParams)
 
 	noOps := uint32(len(*pkOperations))
