@@ -1,6 +1,6 @@
 /*
  * This file is part of the RonDB REST API Server
- * Copyright (c) 2022 Hopsworks AB
+ * Copyright (c) 2023 Hopsworks AB
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	"hopsworks.ai/rdrs/internal/config"
-	"hopsworks.ai/rdrs/internal/integrationtests"
+	"hopsworks.ai/rdrs/internal/integrationtests/testclient"
 	"hopsworks.ai/rdrs/internal/testutils"
 	"hopsworks.ai/rdrs/pkg/api"
 	"hopsworks.ai/rdrs/resources/testdbs"
@@ -31,42 +31,42 @@ import (
 func TestBatchSimple(t *testing.T) {
 	tests := map[string]api.BatchOperationTestInfo{
 		"simple1": { // single operation batch
-			HttpCode: http.StatusOK,
+			HttpCode: []int{http.StatusOK},
 			Operations: []api.BatchSubOperationTestInfo{
 				{
 					SubOperation: api.BatchSubOp{
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB004 + "/int_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", 0, "id1", 0),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", 0, "id1", 0),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "int_table",
 					DB:       testdbs.DB004,
-					HttpCode: http.StatusOK,
+					HttpCode: []int{http.StatusOK},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 			},
 			ErrMsgContains: "",
 		},
 		"simple2": { // small batch operation
-			HttpCode: http.StatusOK,
+			HttpCode: []int{http.StatusOK},
 			Operations: []api.BatchSubOperationTestInfo{
 				{
 					SubOperation: api.BatchSubOp{
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB004 + "/int_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", 0, "id1", 0),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", 0, "id1", 0),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "int_table",
 					DB:       testdbs.DB004,
-					HttpCode: http.StatusOK,
+					HttpCode: []int{http.StatusOK},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 				{
@@ -74,35 +74,35 @@ func TestBatchSimple(t *testing.T) {
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB005 + "/bigint_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", 0, "id1", 0),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", 0, "id1", 0),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "bigint_table",
 					DB:       testdbs.DB005,
-					HttpCode: http.StatusOK,
+					HttpCode: []int{http.StatusOK},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 			},
 			ErrMsgContains: "",
 		},
 		"simple3": { // bigger batch of numbers table
-			HttpCode: http.StatusOK,
+			HttpCode: []int{http.StatusOK},
 			Operations: []api.BatchSubOperationTestInfo{
 				{
 					SubOperation: api.BatchSubOp{
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB004 + "/int_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", 0, "id1", 0),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", 0, "id1", 0),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "int_table",
 					DB:       testdbs.DB004,
-					HttpCode: http.StatusOK,
+					HttpCode: []int{http.StatusOK},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 				{
@@ -110,14 +110,14 @@ func TestBatchSimple(t *testing.T) {
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB005 + "/bigint_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", 0, "id1", 0),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", 0, "id1", 0),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "bigint_table",
 					DB:       testdbs.DB005,
-					HttpCode: http.StatusOK,
+					HttpCode: []int{http.StatusOK},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 				{
@@ -125,14 +125,14 @@ func TestBatchSimple(t *testing.T) {
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB006 + "/tinyint_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", -128, "id1", 0),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", -128, "id1", 0),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "tinyint_table",
 					DB:       testdbs.DB006,
-					HttpCode: http.StatusOK,
+					HttpCode: []int{http.StatusOK},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 				{
@@ -140,14 +140,14 @@ func TestBatchSimple(t *testing.T) {
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB007 + "/smallint_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", 32767, "id1", 65535),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", 32767, "id1", 65535),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "smallint_table",
 					DB:       testdbs.DB007,
-					HttpCode: http.StatusOK,
+					HttpCode: []int{http.StatusOK},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 				{
@@ -155,35 +155,35 @@ func TestBatchSimple(t *testing.T) {
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB007 + "/smallint_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", 1, "id1", 1),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", 1, "id1", 1),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "smallint_table",
 					DB:       testdbs.DB007,
-					HttpCode: http.StatusOK,
+					HttpCode: []int{http.StatusOK},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 			},
 			ErrMsgContains: "",
 		},
 		"notfound": { // a batch operation with operations throwing 404
-			HttpCode: http.StatusOK,
+			HttpCode: []int{http.StatusOK},
 			Operations: []api.BatchSubOperationTestInfo{
 				{
 					SubOperation: api.BatchSubOp{
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB004 + "/int_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", 100, "id1", 100),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", 100, "id1", 100),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "int_table",
 					DB:       testdbs.DB004,
-					HttpCode: http.StatusNotFound,
+					HttpCode: []int{http.StatusNotFound},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 				{
@@ -191,27 +191,27 @@ func TestBatchSimple(t *testing.T) {
 						Method:      &[]string{config.PK_HTTP_VERB}[0],
 						RelativeURL: &[]string{string(testdbs.DB005 + "/bigint_table/" + config.PK_DB_OPERATION)}[0],
 						Body: &api.PKReadBody{
-							Filters:     integrationtests.NewFiltersKVs("id0", 100, "id1", 100),
-							ReadColumns: integrationtests.NewReadColumns("col", 2),
-							OperationID: integrationtests.NewOperationID(64),
+							Filters:     testclient.NewFiltersKVs("id0", 100, "id1", 100),
+							ReadColumns: testclient.NewReadColumns("col", 2),
+							OperationID: testclient.NewOperationID(64),
 						},
 					},
 					Table:    "bigint_table",
 					DB:       testdbs.DB005,
-					HttpCode: http.StatusNotFound,
+					HttpCode: []int{http.StatusNotFound},
 					RespKVs:  []interface{}{"col0", "col1"},
 				},
 			},
 			ErrMsgContains: "",
 		},
 	}
-	integrationtests.BatchTest(t, tests, false)
+	batchTestMultiple(t, tests, false, true)
 }
 
 func TestBatchDate(t *testing.T) {
 	tests := map[string]api.BatchOperationTestInfo{
 		"date": {
-			HttpCode: http.StatusOK,
+			HttpCode: []int{http.StatusOK},
 			Operations: []api.BatchSubOperationTestInfo{
 				createSubOperation(t, "date_table", testdbs.DB019, "1111-11-11", http.StatusOK),
 				createSubOperation(t, "date_table", testdbs.DB019, "1111-11-11 00:00:00", http.StatusOK),
@@ -219,7 +219,7 @@ func TestBatchDate(t *testing.T) {
 			},
 		},
 		"wrong_sub_op": {
-			HttpCode: http.StatusBadRequest,
+			HttpCode: []int{http.StatusBadRequest},
 			Operations: []api.BatchSubOperationTestInfo{
 				createSubOperation(t, "date_table", testdbs.DB019, "1111-11-11", http.StatusOK),
 				createSubOperation(t, "date_table", testdbs.DB019, "1111-11-11 00:00:00", http.StatusOK),
@@ -228,13 +228,13 @@ func TestBatchDate(t *testing.T) {
 			},
 		},
 	}
-	integrationtests.BatchTest(t, tests, false)
+	batchTestMultiple(t, tests, false, true)
 }
 
 func TestBatchDateTime(t *testing.T) {
 	tests := map[string]api.BatchOperationTestInfo{
 		"date": {
-			HttpCode: http.StatusOK,
+			HttpCode: []int{http.StatusOK},
 			Operations: []api.BatchSubOperationTestInfo{
 				createSubOperation(t, "date_table0", testdbs.DB020, "1111-11-11 11:11:11", http.StatusOK),
 				createSubOperation(t, "date_table3", testdbs.DB020, "1111-11-11 11:11:11.123", http.StatusOK),
@@ -249,7 +249,7 @@ func TestBatchDateTime(t *testing.T) {
 			ErrMsgContains: "",
 		},
 		"wrong_sub_op": {
-			HttpCode: http.StatusBadRequest,
+			HttpCode: []int{http.StatusBadRequest},
 			Operations: []api.BatchSubOperationTestInfo{
 				createSubOperation(t, "date_table0", testdbs.DB020, "1111-11-11 11:11:11", http.StatusOK),
 				createSubOperation(t, "date_table3", testdbs.DB020, "1111-11-11 11:11:11.123", http.StatusOK),
@@ -265,13 +265,13 @@ func TestBatchDateTime(t *testing.T) {
 			ErrMsgContains: "",
 		},
 	}
-	integrationtests.BatchTest(t, tests, false)
+	batchTestMultiple(t, tests, false, true)
 }
 
 func TestBatchTime(t *testing.T) {
 	tests := map[string]api.BatchOperationTestInfo{
 		"date": {
-			HttpCode: http.StatusOK,
+			HttpCode: []int{http.StatusOK},
 			Operations: []api.BatchSubOperationTestInfo{
 				createSubOperation(t, "time_table0", testdbs.DB021, "11:11:11", http.StatusOK),
 				createSubOperation(t, "time_table3", testdbs.DB021, "11:11:11.123", http.StatusOK),
@@ -285,7 +285,7 @@ func TestBatchTime(t *testing.T) {
 			ErrMsgContains: "",
 		},
 		"wrong_sub_op": {
-			HttpCode: http.StatusBadRequest,
+			HttpCode: []int{http.StatusBadRequest},
 			Operations: []api.BatchSubOperationTestInfo{
 				createSubOperation(t, "time_table0", testdbs.DB021, "11:11:11", http.StatusOK),
 				createSubOperation(t, "time_table3", testdbs.DB021, "11:11:11.123", http.StatusOK),
@@ -300,24 +300,24 @@ func TestBatchTime(t *testing.T) {
 			ErrMsgContains: "",
 		},
 	}
-	integrationtests.BatchTest(t, tests, false)
+	batchTestMultiple(t, tests, false, true)
 }
 
-func createSubOperation(t *testing.T, table string, database string, pk string, expectedStatus int) api.BatchSubOperationTestInfo {
+func createSubOperation(t testing.TB, table string, database string, pk string, expectedStatus int) api.BatchSubOperationTestInfo {
 	respKVs := []interface{}{"col0"}
 	return api.BatchSubOperationTestInfo{
 		SubOperation: api.BatchSubOp{
 			Method:      &[]string{config.PK_HTTP_VERB}[0],
 			RelativeURL: &[]string{string(database + "/" + table + "/" + config.PK_DB_OPERATION)}[0],
 			Body: &api.PKReadBody{
-				Filters:     integrationtests.NewFiltersKVs("id0", pk),
-				ReadColumns: integrationtests.NewReadColumns("col", 1),
-				OperationID: integrationtests.NewOperationID(5),
+				Filters:     testclient.NewFiltersKVs("id0", pk),
+				ReadColumns: testclient.NewReadColumns("col", 1),
+				OperationID: testclient.NewOperationID(5),
 			},
 		},
 		Table:    table,
 		DB:       database,
-		HttpCode: expectedStatus,
+		HttpCode: []int{expectedStatus},
 		RespKVs:  respKVs,
 	}
 }
@@ -349,7 +349,7 @@ func TestBatchArrayTableLongVarbinary(t *testing.T) {
 func ArrayColumnBatchTest(t *testing.T, table string, database string, isBinary bool, colWidth int, padding bool) {
 	tests := map[string]api.BatchOperationTestInfo{
 		"simple1": { // bigger batch of array column table
-			HttpCode: http.StatusOK,
+			HttpCode: []int{http.StatusOK},
 			Operations: []api.BatchSubOperationTestInfo{
 				arrayColumnBatchTestSubOp(t, table, database, isBinary, colWidth, padding, "-1", http.StatusNotFound),
 				arrayColumnBatchTestSubOp(t, table, database, isBinary, colWidth, padding, "1", http.StatusOK),
@@ -364,11 +364,11 @@ func ArrayColumnBatchTest(t *testing.T, table string, database string, isBinary 
 		},
 	}
 
-	integrationtests.BatchTest(t, tests, isBinary)
+	batchTestMultiple(t, tests, isBinary, true)
 }
 
 /*
- A bad sub operation fails the entire batch
+A bad sub operation fails the entire batch
 */
 func TestBatchBadSubOp(t *testing.T) {
 	table := "table1"
@@ -379,11 +379,11 @@ func TestBatchBadSubOp(t *testing.T) {
 
 	tests := map[string]api.BatchOperationTestInfo{
 		"simple1": { // bigger batch of array column table
-			HttpCode: http.StatusBadRequest,
+			HttpCode: []int{http.StatusBadRequest},
 			Operations: []api.BatchSubOperationTestInfo{
 				arrayColumnBatchTestSubOp(t, table, database, isBinary, colWidth, padding, "-1", http.StatusNotFound),
 				// This is bad operation. data is longer than the column width
-				arrayColumnBatchTestSubOp(t, table, database, isBinary, colWidth, padding, *integrationtests.NewOperationID(colWidth*4 + 1), http.StatusNotFound),
+				arrayColumnBatchTestSubOp(t, table, database, isBinary, colWidth, padding, *testclient.NewOperationID(colWidth*4 + 1), http.StatusNotFound),
 				arrayColumnBatchTestSubOp(t, table, database, isBinary, colWidth, padding, "1", http.StatusOK),
 				arrayColumnBatchTestSubOp(t, table, database, isBinary, colWidth, padding, "2", http.StatusOK),
 				arrayColumnBatchTestSubOp(t, table, database, isBinary, colWidth, padding, "3", http.StatusOK),
@@ -392,7 +392,7 @@ func TestBatchBadSubOp(t *testing.T) {
 		},
 	}
 
-	integrationtests.BatchTest(t, tests, isBinary)
+	batchTestMultiple(t, tests, isBinary, true)
 }
 
 func arrayColumnBatchTestSubOp(t *testing.T, table string, database string, isBinary bool, colWidth int, padding bool, pk string, expectedStatus int) api.BatchSubOperationTestInfo {
@@ -402,14 +402,14 @@ func arrayColumnBatchTestSubOp(t *testing.T, table string, database string, isBi
 			Method:      &[]string{config.PK_HTTP_VERB}[0],
 			RelativeURL: &[]string{string(database + "/" + table + "/" + config.PK_DB_OPERATION)}[0],
 			Body: &api.PKReadBody{
-				Filters:     integrationtests.NewFiltersKVs("id0", integrationtests.Encode(pk, isBinary, colWidth, padding)),
-				ReadColumns: integrationtests.NewReadColumns("col", 1),
-				OperationID: integrationtests.NewOperationID(5),
+				Filters:     testclient.NewFiltersKVs("id0", testclient.EncodePkValue(pk, isBinary, colWidth, padding)),
+				ReadColumns: testclient.NewReadColumns("col", 1),
+				OperationID: testclient.NewOperationID(5),
 			},
 		},
 		Table:    table,
 		DB:       database,
-		HttpCode: expectedStatus,
+		HttpCode: []int{expectedStatus},
 		RespKVs:  respKVs,
 	}
 }
@@ -421,32 +421,32 @@ func TestBatchMissingReqField(t *testing.T) {
 	operations[1].Method = nil
 	operationsWrapper := api.BatchOpRequest{Operations: &operations}
 	body, _ := json.Marshal(operationsWrapper)
-	integrationtests.SendHttpRequest(t, config.BATCH_HTTP_VERB, url, string(body), http.StatusBadRequest,
-		"Error:Field validation for 'Method' failed ")
+	testclient.SendHttpRequest(t, config.BATCH_HTTP_VERB, url, string(body),
+		"Error:Field validation for 'Method' failed ", http.StatusBadRequest)
 
 	// Test missing relative URL
 	operations = NewOperationsTBD(t, 3)
 	operations[1].RelativeURL = nil
 	operationsWrapper = api.BatchOpRequest{Operations: &operations}
 	body, _ = json.Marshal(operationsWrapper)
-	integrationtests.SendHttpRequest(t, config.BATCH_HTTP_VERB, url, string(body), http.StatusBadRequest,
-		"Error:Field validation for 'RelativeURL' failed ")
+	testclient.SendHttpRequest(t, config.BATCH_HTTP_VERB, url, string(body),
+		"Error:Field validation for 'RelativeURL' failed ", http.StatusBadRequest)
 
 	// Test missing body
 	operations = NewOperationsTBD(t, 3)
 	operations[1].Body = nil
 	operationsWrapper = api.BatchOpRequest{Operations: &operations}
 	body, _ = json.Marshal(operationsWrapper)
-	integrationtests.SendHttpRequest(t, config.BATCH_HTTP_VERB, url, string(body), http.StatusBadRequest,
-		"Error:Field validation for 'Body' failed ")
+	testclient.SendHttpRequest(t, config.BATCH_HTTP_VERB, url, string(body),
+		"Error:Field validation for 'Body' failed ", http.StatusBadRequest)
 
 	// Test missing filter in an operation
 	operations = NewOperationsTBD(t, 3)
-	*&operations[1].Body.Filters = nil
+	operations[1].Body.Filters = nil
 	operationsWrapper = api.BatchOpRequest{Operations: &operations}
 	body, _ = json.Marshal(operationsWrapper)
-	integrationtests.SendHttpRequest(t, config.BATCH_HTTP_VERB, url, string(body), http.StatusBadRequest,
-		"Error:Field validation for 'Filters' failed")
+	testclient.SendHttpRequest(t, config.BATCH_HTTP_VERB, url, string(body),
+		"Error:Field validation for 'Filters' failed", http.StatusBadRequest)
 }
 
 func NewOperationsTBD(t *testing.T, numOps int) []api.BatchSubOp {
@@ -458,7 +458,7 @@ func NewOperationsTBD(t *testing.T, numOps int) []api.BatchSubOp {
 }
 
 func NewOperationTBD(t *testing.T) api.BatchSubOp {
-	pkOp := integrationtests.NewPKReadReqBodyTBD()
+	pkOp := testclient.NewPKReadReqBodyTBD()
 	method := "POST"
 	relativeURL := testutils.NewPKReadURL("db", "table")
 
