@@ -1,3 +1,20 @@
+/*
+ * This file is part of the RonDB REST API Server
+ * Copyright (c) 2023 Hopsworks AB
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package rest
 
 import (
@@ -20,7 +37,7 @@ func (h *RouteHandler) PkRead(c *gin.Context) {
 	var responseIntf api.PKReadResponse = (api.PKReadResponse)(&api.PKReadResponseJSON{})
 	responseIntf.Init()
 
-	status, err := handlers.Handle(h.pkReadHandler, &apiKey, pkReadParams, responseIntf)
+	status, err := handlers.Handle(&h.pkReadHandler, &apiKey, pkReadParams, responseIntf)
 	if err != nil {
 		c.AbortWithError(status, err)
 		return
@@ -29,7 +46,7 @@ func (h *RouteHandler) PkRead(c *gin.Context) {
 }
 
 /*
-	Parsing both GET and POST parameters
+Parsing both GET and POST parameters
 */
 func parsePkReadRequest(c *gin.Context) (*api.PKReadParams, error) {
 	getParams := api.PKReadPP{}
