@@ -1,3 +1,20 @@
+/*
+ * This file is part of the RonDB REST API Server
+ * Copyright (c) 2023 Hopsworks AB
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package rest
 
 import (
@@ -12,7 +29,7 @@ import (
 func (h *RouteHandler) Stat(c *gin.Context) {
 	apiKey := c.GetHeader(config.API_KEY_NAME)
 	statResp := api.StatResponse{}
-	status, err := handlers.Handle(h.statsHandler, &apiKey, nil, &statResp)
+	status, err := handlers.Handle(&h.statsHandler, &apiKey, nil, &statResp)
 	if err != nil {
 		c.AbortWithError(http.StatusInternalServerError, err)
 		return
