@@ -1,3 +1,20 @@
+/*
+ * This file is part of the RonDB REST API Server
+ * Copyright (c) 2023 Hopsworks AB
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package servers
 
 import (
@@ -10,21 +27,24 @@ import (
 	"hopsworks.ai/rdrs/internal/dal/heap"
 	"hopsworks.ai/rdrs/internal/log"
 
-	"hopsworks.ai/rdrs/internal/security/apikey/authcache"
+	"hopsworks.ai/rdrs/internal/security/apikey"
 	"hopsworks.ai/rdrs/internal/security/tlsutils"
 	"hopsworks.ai/rdrs/internal/servers/grpc"
 	"hopsworks.ai/rdrs/internal/servers/rest"
 )
 
-func CreateAndStartDefaultServers(heap *heap.Heap, quit chan os.Signal) (err error, cleanup func()) {
+func CreateAndStartDefaultServers(
+	heap *heap.Heap,
+	apiKeyCache apikey.Cache,
+	quit chan os.Signal,
+) (cleanup func(), err error) {
 	cleanup = func() {}
 
 	// Connect to RonDB
 	conf := config.GetAll()
-	connectString := config.GenerateMgmdConnectString(conf)
-	dalErr := dal.InitRonDBConnection(connectString, true)
+	dalErr := dal.InitRonDBConnection(conf.RonDB)
 	if dalErr != nil {
-		return fmt.Errorf("failed to connect to RonDB %s. error: %w", connectString, dalErr), cleanup
+		return cleanup, fmt.Errorf("failed to connect to RonDB. error: %w", dalErr)
 	}
 	cleanupRonDB := func() {
 		dalErr = dal.ShutdownConnection()
@@ -34,21 +54,21 @@ func CreateAndStartDefaultServers(heap *heap.Heap, quit chan os.Signal) (err err
 	}
 
 	var tlsConfig *tls.Config
-	if conf.Security.EnableTLS {
+	if conf.Security.TLS.EnableTLS {
 		tlsConfig, err = tlsutils.GenerateTLSConfig(
-			conf.Security.RequireAndVerifyClientCert,
-			conf.Security.RootCACertFile,
-			conf.Security.CertificateFile,
-			conf.Security.PrivateKeyFile,
+			conf.Security.TLS.RequireAndVerifyClientCert,
+			conf.Security.TLS.RootCACertFile,
+			conf.Security.TLS.CertificateFile,
+			conf.Security.TLS.PrivateKeyFile,
 		)
 		if err != nil {
 			cleanupRonDB()
-			return fmt.Errorf("failed generating tls configuration; error: %w", err), cleanup
+			return cleanup, fmt.Errorf("failed generating tls configuration; error: %w", err)
 		}
 	}
 
-	grpcServer := grpc.New(tlsConfig, heap)
-	err, cleanupGrpc := grpc.Start(
+	grpcServer := grpc.New(tlsConfig, heap, apiKeyCache)
+	cleanupGrpc, err := grpc.Start(
 		grpcServer,
 		conf.GRPC.ServerIP,
 		conf.GRPC.ServerPort,
@@ -56,7 +76,7 @@ func CreateAndStartDefaultServers(heap *heap.Heap, quit chan os.Signal) (err err
 	)
 	if err != nil {
 		cleanupRonDB()
-		return fmt.Errorf("failed starting gRPC server; error: %w", err), cleanup
+		return cleanup, fmt.Errorf("failed starting gRPC server; error: %w", err)
 	}
 
 	restServer := rest.New(
@@ -64,14 +84,13 @@ func CreateAndStartDefaultServers(heap *heap.Heap, quit chan os.Signal) (err err
 		conf.REST.ServerPort,
 		tlsConfig,
 		heap,
+		apiKeyCache,
 	)
+
 	cleanupRest := restServer.Start(quit)
-	return nil, func() {
+	return func() {
 		cleanupRonDB()
 		cleanupGrpc()
 		cleanupRest()
-
-		// Clean API Key Cache
-		authcache.Reset()
-	}
+	}, nil
 }
