@@ -1,3 +1,20 @@
+/*
+ * This file is part of the RonDB REST API Server
+ * Copyright (c) 2023 Hopsworks AB
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package testutils
 
 import (
@@ -60,7 +77,7 @@ func NewBatchReadURL() string {
 
 func appendURLProtocol(url *string) {
 	conf := config.GetAll()
-	if conf.Security.EnableTLS {
+	if conf.Security.TLS.EnableTLS {
 		*url = fmt.Sprintf("https://%s", *url)
 	} else {
 		*url = fmt.Sprintf("http://%s", *url)
