@@ -1,6 +1,6 @@
 /*
  * This file is part of the RonDB REST API Server
- * Copyright (c) 2022 Hopsworks AB
+ * Copyright (c) 2023 Hopsworks AB
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,6 +22,7 @@ package api
 */
 import "C"
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
 	"strings"
@@ -97,6 +98,7 @@ type PKReadResponse interface {
 	Init()
 	SetOperationID(opID *string)
 	SetColumnData(column, value *string, valueType uint32)
+	String() string
 }
 
 type PKReadResponseJSON struct {
@@ -126,6 +128,22 @@ func (r *PKReadResponseGRPC) SetColumnData(column, value *string, valueType uint
 	}
 }
 
+func (r *PKReadResponseGRPC) String() string {
+	var str bytes.Buffer
+	str.WriteString("{ ")
+	str.WriteString(fmt.Sprintf("\"OperationID\": \"%s\",", *r.OperationID))
+	str.WriteString("\"Data\": {")
+
+	if r.Data != nil {
+		for key, value := range *r.Data {
+			str.WriteString(fmt.Sprintf("\"%s\": \"%s\",", key, *value))
+		}
+	}
+
+	str.WriteString("}")
+	return str.String()
+}
+
 func (r *PKReadResponseJSON) Init() {
 	m := make(map[string]*json.RawMessage)
 	(*r).Data = &m
@@ -150,13 +168,20 @@ func (r *PKReadResponseJSON) SetColumnData(column, value *string, dataType uint3
 	}
 }
 
-var _ PKReadResponse = (*PKReadResponseGRPC)(nil)
-var _ PKReadResponse = (*PKReadResponseJSON)(nil)
+func (r *PKReadResponseJSON) String() string {
+	strBytes, err := json.MarshalIndent(*r, "", "\t")
+	if err != nil {
+		return fmt.Sprintf("Failed to marshar PKReadResponseJSON. Error: %v", err)
+	} else {
+		return string(strBytes)
+	}
+}
 
 type PKReadResponseWithCode interface {
 	Init()
 	GetPKReadResponse() PKReadResponse
 	SetCode(code *int32)
+	String() string
 }
 
 type PKReadResponseWithCodeJSON struct {
@@ -182,6 +207,15 @@ func (p *PKReadResponseWithCodeJSON) SetCode(code *int32) {
 	p.Code = code
 }
 
+func (p *PKReadResponseWithCodeJSON) String() string {
+	strBytes, err := json.MarshalIndent(*p, "", "\t")
+	if err != nil {
+		return fmt.Sprintf("Failed to marshar PKReadResponseJSON. Error: %v", err)
+	} else {
+		return string(strBytes)
+	}
+}
+
 func (p *PKReadResponseWithCodeGRPC) Init() {
 	p.Body = &PKReadResponseGRPC{}
 	p.Body.Init()
@@ -195,8 +229,20 @@ func (p *PKReadResponseWithCodeGRPC) SetCode(code *int32) {
 	p.Code = code
 }
 
-var _ PKReadResponseWithCode = (*PKReadResponseWithCodeJSON)(nil)
-var _ PKReadResponseWithCode = (*PKReadResponseWithCodeGRPC)(nil)
+func (p *PKReadResponseWithCodeGRPC) String() string {
+	var str bytes.Buffer
+	str.WriteString("{ ")
+	str.WriteString(fmt.Sprintf("\"Code\": \"%d\",", *p.Code))
+
+	str.WriteString("\"Body\": { ")
+	if p.Body != nil {
+		str.WriteString(p.Body.String())
+	}
+	str.WriteString("} ") // Body
+
+	str.WriteString("}")
+	return str.String()
+}
 
 // For testing only
 type PKTestInfo struct {
