@@ -1,6 +1,6 @@
 //
 // This file is part of the RonDB REST API Server
-// Copyright (c) 2022 Hopsworks AB
+// Copyright (c) 2023 Hopsworks AB
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
